Wrap up large filenames in CLI output
