+++ src/render/cliRenderer/utils.js
@@ -7,4 +7,13 @@ function enableKeyPressEvent() {
   process.stdin.resume();
 }
 
+function wrapUp(string, maxLength) {
+  if (string.length <= maxLength) {
+    return string;
+  }
+  const regex = new RegExp(`.{1,${maxLength}}`, 'g');
+  return string.match(regex).join('\n');
+}
+
 exports.enableKeyPressEvent = enableKeyPressEvent;
+exports.wrapUp = wrapUp;
